Refactor filestructure
